Corrected from_response (#5)
This PR intends to correct the from_response method by not directly calling new with response but formatting the amount and then calling new with the attributes. The afterpay-sdk gem follows the Data Models attibutes data types as in v2 doc. As pe the doc amount attibute of Refund or PaymentEvent object is a Money object so when the developer calls new he need to supply the money object however we use from_response t method where we convert amount in response to money object using utils.
https://yourmechanic.atlassian.net/browse/PLAT-351